backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
paramiko==3.5.1
cryptography==45.0.4
pgvector==0.4.1
ase==3.24.0
